+++ unicorn_binance_websocket_api/unicorn_binance_websocket_api_connection.py
@@ -252,5 +252,20 @@ class BinanceWebSocketApiConnection(object):
         try:
             await self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].send(data)
             self.handler_binance_websocket_api_manager.increase_transmitted_counter(self.stream_id)
+        except websockets.exceptions.ConnectionClosed as error_msg:
+            logging.error("BinanceWebSocketApiSocket->send(" + str(self.stream_id) + ", " +
+                          str(self.channels) + ", " + str(self.markets) + ") Exception ConnectionClosed "
+                          "Info: " + str(error_msg))
+            if "WebSocket connection is closed: code = 1006" in str(error_msg):
+                self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].close()
+                self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
+                sys.exit(1)
+        except RuntimeError as error_msg:
+            logging.error("BinanceWebSocketApiSocket->send(" + str(self.stream_id) + ", " +
+                          str(self.channels) + ", " + str(self.markets) + ") Exception RuntimeError "
+                          "Info: " + str(error_msg))
+            self.handler_binance_websocket_api_manager.websocket_list[self.stream_id].close()
+            self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(error_msg))
+            sys.exit(1)
         except KeyError:
             pass